Add database to repository
The database was not originally included in the repository to prevent
public access to to user data. When this policy was accepted, the
development team was not aware that doing a migration on a computer
changes that computer's copy of the database. When the developers would
do a migration locally and push their changes to the remote repository,
the uploaded migration files would reflect the migrations they performed
on their computer. Since the virtual machine pulls from the remote
repository, the VM's migration files would reflect all the migrations
done on the developers' computers. However, since the VM's database did
not track the developer's database, the VM database would not reflect
the migrations done on the developers' computers. This conflict between
the migration files and the database on the VM prevented any migrations
-- whether done on the VM via SSH or on a developer's computer -- from
effecting the VM's copy of the Django project.

This commit adds the database to the remote repository so that it is
synchronized across all of the developer's computers and the VM. This is
not an ideal solution since it essentially makes all the user data
publicly accessible. Though passwords are hashed, we store enough
sensitive data that security should be a primary concern. This solution
is absolutely development-only. We will need to research other database
solutions for production.